requirements.txt
crawl4AI==0.4.247
python-dotenv==1.0.1
pydantic==2.10.6
transformers 
torch
nltk
openai
sentence-transformers
chromadb
langchain-community
streamlit
langchain